Reorganize early-stage docs sections (#19603)
Backports #19205

* Reorganize early-stage docs sections

Closes #18127

Currently, there are two sections of the docs related to the earliest
stages of setting up Teleport, "Try out Teleport" and "Deploy a
Cluster". While the organization of our top-level docs sections attempts
to correspond to the stages a user goes through when setting up
Teleport, these combine some tasks that should probably be separated:

- **Choosing an edition and deploying a production cluster**: While the
  "Deploy a Cluster" section includes our deployment guides as well as
  our edition guides, user will likely commit to setting up Teleport
  Cloud, Open Source, or Enterprise before they plan a full HA
  production deployment.

- **Deploying a production cluster and deploying a small-scale Teleport
  cluster:** While the "Deploy a Cluster" section includes the Linux
  Server and Digital Ocean 1-Click guides, these two guides are more for
  small-scale deployments than full production clusters.

This change improves the early-stage docs experience for new Teleport
users by:

Splitting the "Deploy a Cluster" section
----------------------------------------

The "Deploy a Cluster" section includes subsections related to choosing
a Teleport edition as well as subsections related to deploying a
production cluster. The problem with this approach is that it leaves no
room for adding introductory guides to help users with these two
purposes.

By splitting the "Deploy a Cluster" section into two top-level sections,
one related to choosing an edition and the other to dpeloying to
production, we can add introductory pages that put the subsections of
these sections in context.

_This change accompanies the work on #16751 to write a general guide to
HA Teleport deployments, since the new guide will have a prominent
place in the reorganized "Deploy a Cluster" section._

Adding introductory pages
-------------------------

Sections related to getting started with Teleport, "Try out Teleport"
and "Deploy a Cluster", were missing introductory pages. This change
adds them, and also adds one to the new "Choose an Edition" section.

Moving the Linux Server and Digital Ocean guides
------------------------------------------------

Move these to "Try out Teleport". The Digital Ocean guide didn't fit
with the rest of the deployment guides, since it's not about
architecting an HA deployment.

This is also a better home for the "Linux Server" guide, so this change
moves that as well.

Removing the "getting-started.mdx" page
---------------------------------------

This page is not reachable via the navigation menu.

* Respond to PR feedback

* Respond to PR feedback

Also minor text tweaks