Rework plugin service to user CreatePluginRequest
The intent is to have
the auth server exchange initial credentials for live credentials,
as opposed to the previous approach, where
it is the individual plugin instances that do the exchange.

Since the initial credentials are truly one-time-use,
the previous approach is hard to implement correctly in an HA environment.