DatabaseService Heartbeat: return AWS Identity
Why?
In the context of Teleport Discover, we want to ease the process of
adding resources.
When adding an AWS managed Database, it's useful to know the current AWS
Identity and the Account ID where the DatabaseService is running.
This allow us to create the required policy to enable proxying the
Database.
This policy is then presented to the user, with a simple guide on how to
add it.
Creating the policy requires some fields, and when we don't have them,
we show placeholders.
Those placeholders are not always very clear or very user friendly.

To ease this process, we can fetch the required fields if we already
have a DatabaseService running in a EC2 instance.

How?
When a DatabaseService starts, it emits hearbeats to the cluster
providing keep alive hearbeats but also some other information like the
resource matchers that this Service is monitoring.

With this PR, there's some other information sent back to the cluster:
- Instance Metadata - AWS Identity
Which is then used to create the required policy as described above.