Update watched resources for RemoteOldProxy cache
Update the version we should check against: v12 (the current in master)

Update the list of resources to watch for RemoteOldProxy.
The list should be equal to the RemoteProxy as seen in `branch/v11`

As soon as we cut v12, we should update this again in master because
we'll be preparing v13.