Stablize RemoteConnCleanup (#20048) (#20090)
Alters `RemoteConnCleanup` to rely less on `BlockUntil` by artificially
adjusting the last heartbeat manually in the test. Additionally, the
tests in `localsite_test` no longer call `native.PreComputeKey` by
injecting an uninitialized `certificateCache` into the localsite
constructor since it is not used for the tests. The logger is now
also initialized for all tests within the `reversetunnel` package.

Closes #19827