Add --quiet to eslint package.json script (#1510) (#1523)
By default, eslint outputs warnings as well as errors but doesn't sort
them in any way. When the lint step fails in CI because of an error, it's
hard to find that error among warnings.

The warnings are still useful when they show up in the editor, so let's
add --quiet which makes eslint output only errors and no warnings.