WebAPI/CreateDB: improve error message when DB already exists
When creating a database with the same name of an already existing one,
we were getting a raw error (depends on the backend implementation:
memory, lite, dynamo, ...).
This message is shown to the user which sees it as a cryptic error.

To improve this we should present the error in a better way and provide
an immediate action.