Fix kube access proxy peering compatibility
The change from kube_service to kube_server in v11
lead to breaking backwards compatibility for v10 agents
connecting to a v11 teleport cluster when proxy peering
is enabled.

The issue was in converting from a kube_service to kube_server
the proxy ids the kube agent is connected to was never copied.

This leads to kube agents being reachable through the proxy
they are connected to but not through peer proxies.