Introduce a "UserDomain" configuration property
The existing "domain" field in the LDAP configuration is now used
for the domain where we publish the Teleport CA and where we do
desktop discovery.

If the user's you wish to generate credentials for are in the same
domain as the desktops, this change is transparent. In cases where
you want to generate user credentials for a different domain, you
can override the domain by setting this new field.

Notes:

1. The user domain is also used to generate Teleport's own
   service account credentials.
2. SID-mapping is temporarily disabled, as the SID lookup would
   need to be performed against a different LDAP server when the
   user domain is overridden.