Set certain ssh env variables without waiting for a reply
Setting an env on an ssh session requires waiting for the server
to process and reply to the request. This makes setting N env vars
in a row a rather slow task as they aren't done in parallel. Since
we only log the returned error and continue on if setting an env
fails in some cases we can get a slight reduction in latency if
we don't request a reply from the server.