This PR introduces a new ALPN protocol extension forwards proxy connections to a gRPC server with mTLS authentication. (#20808)
Proxy already has a `gRPC` server when using the `teleport-proxy-grpc` ALPN protocol. Still, that server does not impose mTLS auth since it's used by nodes when joining the cluster with dynamic joining methods.

Since we require that the user identity is available in the handler, this PR introduces a new gRPC server that imposes mTLS under the ALPN protocol `teleport-proxy-gprc-mtls`.

Part of #20324
Related to #19573