helm: Expose ports on containers (#20485)
* helm: Expose ports on containers

Some Kubernetes tooling expects to see the ports that a container exposes registered on the individual containers themselves as well as the underlying Service. This change makes it a little easier for users to deploy the teleport-cluster chart behind their own load balancers or using a ClusterIP, by being explicit about exactly which ports are expected to be routed to each container.

* Expose tls port 3080

* Expose port 3080 before others

* Rename https -> tls

* Add 3021 for proxy peering when enterprise is true