Attempt to reconnec a closed LDAP connection
When desktop discovery via LDAP is enabled, there is enough traffic
on the connection for it to remain open for long periods of time.
If discovery is disabled and there are not frequent connections to
desktops, the LDAP server may close the connection.

In this case, future connection attempts will fail when making an
LDAP query for the user's SID. Now we detect closed connections in
this code path and attempt to open a new LDAP connection.

The first connection attempt after the LDAP connection is closed will
incur some extra latency, as we obtain a new cert when refershing the
LDAP connection.

Additionally, increase the heartbeat period for static hosts from 5s
to 5m. These hosts are infrequently changing, so heartbeating every
5 seconds is an inefficient use of resources.