Fix forwarding of impersonated headers for `kubectl exec/portforward`
This PR fixes the impersonation headers propagation when using the SPDY
roundtrip. SPDY roundtrip creates a new http request and loses the
impersonation headers received from the client. When this happens for
users that define multiple `kubernetes_users`, the requests are denied
because Teleport forces you to select a user.

With the changes introduced we conserve the headers received from the
client.

Fixes #21088