Removes Drone concurrency limit on native windows build (#21201)
Windows builds are currently limited to 1 concurrent build for all stages
with the same name. This is causing unnecessary delay on the push build
queue, given that the Windows native builds are all set up to be independent
of one another.

Also updates the dronegen process to use a stock drone cli instance
proxied via tsh, to better match modern usage.