Prepare for Dockerfile-teleterm removal
`make protos-up-to-date` and thus `make grpc` is now run as a part of CI.
Thus I cannot just change the buildbox Dockerfile when backporting #20032
as it will use the old buildbox image which won't be able to compile JS
protos.

I extracted relevant Dockerfile changes from the backport branch.
Once this commit lands in v10, I should be able to make CI pass for the
backport PR.