Align the Okta and Auth Connector configuration examples in Okta SSO guide (#21466)
Port `:443` is explicitly stated in the Okta configuration but not in the the Teleport auth connector example (only has <cluster-url> mentioned without 443).  This has created multiple Okta issues for users recently with mismatched configurations on the auth connector and Okta configuration.

Proposal is to align the configuration reference for the Okta and Teleport auth connector examples so users are clear these MUST match for the connector to work.  I chose to align the Okta to the Teleport example but it could go the other way too.

We may also want to consider changing the order of the Okta guide.  With IDP initiated configuration flow it makes more sense to create and name the Teleport connector first before trying to configure Okta.  But this is outside the scope of this PR for now.