Rename protoEqual and add a big warning
Rename `protoEqual` to `protoKnownFieldsEqual` and add a warning about
its usage.

Manually ignoring `XXX_` fields in full comparisons runs the risk of
ignoring unknown fields, which in itself is a strong hint that we are
running outdated code. The recommended practice for scenarios where full
equality is not desirable is to code it as such, comparing the fields
that matter.

`proto.Equal` should be preferred in most scenarios.