Fix kube agent shutdown during upgrades (#21537)
During binary upgrades, Teleport forks the agents and calls the shutdown
method for each one of them.

Kubernetes Agent was calling the `http.Server` shutdown method and which
wasn't stopping the heartbeats neither handling the listeners.
This PR fixes the shutdown behaviour for Kube agent and avoids deleting
the heartbeated resources when the agent is reloading.