Remove Auth/Proxy instructions from DB guides (#21685)
Backports #20308

* Remove Auth/Proxy instructions from DB guides

Closes #11538

A number of our Database Service guides instruct the user to install the
Auth Service and the Proxy Service. For users who have already set up
these services, these instructions add friction. For users who have
_not_ set up these services, Teleport has existing instructions that
users can follow instead.

This change edits our Database Access guides to require the user to have
a running Teleport cluster as a prerequisite, in line with how-to guides
in other sections of the docs.

Note that, while the issue this change closes is to add Auth/Proxy
Service setup instructions to _more_ places in the docs, this change
follows our style guide recommendation (#20024) to add steps to the
Prerequisites section if we have them documented elsewhere and they
don't directly pertain to the goal of a how-to guide.

* Respond to PR feedback

Change H3s to H2s

---------

Co-authored-by: Alex Fornuto <alex.fornuto@goteleport.com>