Refactor schemagen
Handle all protobuf parsing and output in the schemagen library so
consumers of the library only need to worry about processing JSON Schema
data.

This way, the plugin for generating the configuration resource reference
only needs to include tests for generating populated templates from a
JSON Schema document, and does not need to deal with
CodeGeneratorRequests or CodeGeneratorResponses.