Refer to tsh apps subcommand
The `tsh app` family of subcommands is aliased to `tsh apps`,
so both invocations work correctly. The command itself is defined
as `tsh apps` so this is what appears in the help message.

Update references to `tsh app` so there isn't confusion when
browsing `tsh help` and looking for a missing `tsh app` subcommand.

Backports #21431