Switch CodeQL to scheduled (#882)
This switch mirrors the one recently submitted to `teleport`: https://github.com/gravitational/teleport/pull/21883

Although we are not having timeout issues in .e, it is the third most costly repo from CodeQL executions.  For that reason I think it makes sense to keep it aligned with the `teleport` repo until we figure out a faster option.

Source-commit: gravitational/teleport.e@82681e1ce14ad1ae321a8d31b4baf8e1b8f15e46
Source-pull-requests: gravitational/teleport.e#882