Docs: Remove Jira Custom Field reference
The plugin doesn't write into a custom field like the docs suggest.

The plugin uses the `properties` field of the Create Issue request.
That `properties` field is an array of `Key:string Value:string`.
When the issue is updated in JIRA, that's the value we look for to
update the Teleport Access Request.