Switch CodeQL to scheduled (#21942)
We still plan to address this root issue, and are tracking making CodeQl faster so it can be a required PR check: https://github.com/gravitational/SecOps/issues/269

However, until then we need to switch this job to be a scheduled task rather than on every PR and push.  This is partly cost motivated, but also we are already hitting our timeouts.  This PR also increases the analysis timeout so that the daily job can be sure to complete.