Reduce etcd requests performed by a KeepAlive (#21927)
When performing a KeepAlive the key was first retrieved from
etcd to verify that it existed before updating the item. However,
a Put operation with the `ignore_value` set will return a not
found error if the key does not exist
(https://etcd.io/docs/v3.4/learning/api/#put). So we can safely
remove the get and rely on etcd to prevent updating a key that
does not exist.

Also converts `concurrentqueue.Queue` to be generic over the given
Input and Output of the worker function.