Support proxy reading of SAML IdP CA.
GetCertAuthority was not respecting CA type where clauses because the CA was
not being passed into the AccessChecker properly. This has been fixed.
Additionally, the SAML IdP CA's validation has been fixed so that it supports
more than 1 active key pair during key rotations.

With all this, the proxy will now be able to read the SAML IdP CA.