tsh proxy: pass original remote hostname in ProxyCommand
According to openssh documentation
https://www.man7.org/linux/man-pages/man5/ssh_config.5.html#TOKENS
there is a difference in hostname formatting
when using '%h' or '%n'. This can lead hostnames with uppercases to be
changed to lowercases and the connection to fail.

Rather than change it only on openssh.go,
I choose to change it everywhere it's used because
I can't see
cases where it would be beneficial to have the extra
formatting provided by '%h'.

Using this option requires bumping the minimum
supported openssh version to 8.1 :
https://www.openssh.com/txt/release-8.1

Fixes: https://github.com/gravitational/teleport/issues/16457