Set doc status to connecting when creating doc from connection
This should be changed a long time ago. Currently, there's no other place
which depends on the status being disconnected. That's going to change in
upcoming commits as I'm going to reuse it.

In the meantime, let's change it to 'connecting' so that the progress bar
is correctly shown when creating a new tsh ssh doc based on a connection
from the list.