Enable xterm links and clean up MFA modal
* Clean up the MFA modal (#22045)

Change the text on the MFA dialog to be less alarming.
Additionally, focus the OK button by default, so that users can
press enter to go straight to the MFA prompt without manually
clicking a button.

Closes #19042

* Enable xterm web links addon (#22043)

This renders links as <a> elements in the web-based terminal,
making them true clickable links.

Closes #7569