Automatically parse entity ID from SAML SP during CLI creation. (#22101) (#22368)
* Automatically parse entity ID from SAML SP during CLI creation.

When a SAML IdP service provider is created in the CLI, the entity ID will
be automatically parsed out of the entity descriptor and pushed into the
object before pushing the object to the backend.

Additionally, a small bug was found with the SAML IdP service provider get
logic and fixed.

* Update tool/tctl/common/collection.go



* Update tool/tctl/common/resource_command_test.go



* Bring back copy loop.

* Remove the SkipCheckAndSetDefault.

* Removing the check for the entity ID in the SP object.

* Restore missing comment.

* Parse entity ID as part of check and set defaults.

* Cleanup api go.mod/go.sum.

---------

Co-authored-by: Andrew Burke <31974658+atburke@users.noreply.github.com>