SAML IdP implementation. (#884)
The SAML IdP has been implemented. This allows users to use Teleport as a
SAML identity provider. This implementation adapts some code from
crewjam/saml/samlidp, but ultimately chooses to implement everything
directly as the samlidp implementation doesn't fit our needs directly and is
intended as an example. The following features are implemented by this PR:

* Exposed Metadata endpoint.
* Assertions built off of crewjam/saml's assertion maker.
* Service provider initiated SSO.
* IdP initiated SSO.
* Session handling.
* Handling of service providers.
* Locking of users.

The exposed endpoints are:

* `/enterprise/saml-idp/metadata` for downloading the metadata XML.
* `/enterprise/saml-idp/sso` for service provider initiated SSO.
* `/enterprise/saml-idp/login/:shortcut` for IdP initiated SSO.

The metadata XML must be loaded into a service provider and then the
metadata for the service provider must be loaded into the SAML IdP by
using `tctl create -f saml-idp-service-provider.yaml`.

The assertions are limited to `uid` (the user ID) and `eduPersonAffiliation`
(roles). These can be expanded upon at a later date, but this seemed to be
a reasonable first stab at assertions are users may find interesting.
Additionally, auth requests are able to request a number of attributes
which will be filled in based off of the session value. This block has been
copied verbatim from the samlidp example implementation, so we may want to
expand upon this.

To be implemented:

* Audit events.
* Certificate rotation support.

Source-commit: gravitational/teleport.e@7d5881b0bc0b861811a7931cc60523fecc292b2b
Source-pull-requests: gravitational/teleport.e#884